First CardView Addition for current fragment.

+++ app/src/main/java/com/example/weatherdetailer/network/WeatherResponse.kt
@@ -18,7 +18,7 @@ class WeatherResponse {
     @SerializedName("clouds")
     var clouds:Clouds?= null
     @SerializedName("dt")
-    var dt:Float=0.toFloat()
+    var dt:Long=0
     @SerializedName("id")
     var id:Int=0
     @SerializedName("name")
